+++ packages/excalidraw/components/Actions.tsx
@@ -429,6 +429,7 @@ export const ZoomActions = ({
       {renderAction("zoomOut")}
       {renderAction("resetZoom")}
       {renderAction("zoomIn")}
+      {renderAction("zoomToFit")}
     </Stack.Row>
   </Stack.Col>
 );