+++ packages/excalidraw/components/LayerUI.tsx
@@ -27,7 +27,7 @@ import { FixedSideContainer } from "./FixedSideContainer";
 import { HintViewer } from "./HintViewer";
 import { Island } from "./Island";
 import { LoadingMessage } from "./LoadingMessage";
-import { LockButton } from "./LockButton";
+import { LockElementButton } from "./LockElementButton";
 import { MobileMenu } from "./MobileMenu";
 import { PasteChartDialog } from "./PasteChartDialog";
 import { Section } from "./Section";
@@ -63,6 +63,7 @@ import Scene from "../scene/Scene";
 import { LaserPointerButton } from "./LaserPointerButton";
 import { MagicSettings } from "./MagicSettings";
 import { TTDDialog } from "./TTDDialog/TTDDialog";
+import { actionToggleElementLock } from "../actions";
 
 interface LayerUIProps {
   actionManager: ActionManager;
@@ -234,6 +235,15 @@ const LayerUI = ({
     </Section>
   );
 
+  const allElementsLocked = (elementsIds: string[]) => {
+    if (elementsIds.length === 0) {
+      return false;
+    }
+    return elements
+      .filter((element) => elementsIds.includes(element.id))
+      .every((element) => element.locked);
+  };
+
   const renderFixedSideContainer = () => {
     const shouldRenderSelectedShapeActions = showSelectedShapeActions(
       appState,
@@ -282,10 +292,28 @@ const LayerUI = ({
                             title={t("toolBar.penMode")}
                             penDetected={appState.penDetected}
                           />
+                          {/*
+                            Removed:
                           <LockButton
                             checked={appState.activeTool.locked}
                             onChange={onLockToggle}
                             title={t("toolBar.lock")}
+                          />
+                          */}
+                          <LockElementButton
+                            disabled={
+                              Object.keys(appState.selectedElementIds)
+                                .length === 0
+                            }
+                            checked={allElementsLocked(
+                              Object.keys(appState.selectedElementIds),
+                            )}
+                            onChange={() =>
+                              actionManager.executeAction(
+                                actionToggleElementLock,
+                              )
+                            }
+                            title={t("toolBar.lockElements")}
                           />
 
                           <div className="App-toolbar__divider" />
@@ -347,6 +375,7 @@ const LayerUI = ({
             )}
             {renderTopRightUI?.(device.editor.isMobile, appState)}
             {!appState.viewModeEnabled &&
+              !appState.hideLibraryButton &&
               // hide button when sidebar docked
               (!isSidebarDocked ||
                 appState.openSidebar?.name !== DEFAULT_SIDEBAR.name) && (